Moving page terminals into correct folders
