+++ src/TooMany.WebClient/src/components/monitor/index.ts
@@ -1,1 +1,2 @@
 export * as Header from './header';
+export * as Terminal from './terminal';
