+++ src/TooMany.WebClient/src/components/monitor/terminal/Task.tsx
@@ -1,8 +1,8 @@
 import { Task } from '@hooks/API';
-import Logs from './Logs';
+import Terminal from '@components/terminal';
 
 export default function ({ name }: { name: string }) {
 	const { data: logs = [], isLoading } = Task.log.useLog(name);
 	if (isLoading) return null;
-	return <Logs logs={logs} />;
+	return <Terminal logs={logs} />;
 }
