+++ src/TooMany.WebClient/src/components/monitor/terminal/index.ts
@@ -1,0 +1,2 @@
+export { default as Task } from './Task';
+export { default as Tag } from './Tag';
