+++ src/TooMany.WebClient/src/components/terminal/index.ts
@@ -1,2 +1,1 @@
-export { default as TaskTerminal } from './TaskTerminal';
-export { default as TagTerminal } from './TagTerminal';
+export { default } from './Terminal';
