+++ src/TooMany.WebClient/src/pages/monitor/Task.tsx
@@ -2,17 +2,17 @@ import { memo } from 'react';
 import { useParams } from 'react-router-dom';
 import SuspenseQuery from '@components/helpers/SuspenseQuery';
 import ScrollToBottom from '@components/helpers/ScrollToBottom';
-import { TaskTerminal } from '@components/terminal';
-import { Header } from '@components/monitor';
+import { Task as Header } from '@components/monitor/header';
+import { Task as Terminal } from '@components/monitor/terminal';
 
 function MonitorPage() {
 	const { name } = useParams();
 	return (
 		<section className="max-h-screen flex flex-col">
-			<Header.Task name={name} />
+			<Header name={name} />
 			<SuspenseQuery fallback={<h1>Loading Terminal...</h1>}>
 				<ScrollToBottom>
-					<TaskTerminal name={name} />
+					<Terminal name={name} />
 				</ScrollToBottom>
 			</SuspenseQuery>
 		</section>
